Add image editor to manual
