+++ crates/templates/src/templates/recipes/mod.rs
@@ -1,0 +1,15 @@
+mod add;
+mod add_page;
+mod edit;
+mod index;
+mod search;
+mod view;
+
+pub use add::*;
+pub use add_page::*;
+pub use edit::*;
+pub use index::*;
+pub use search::*;
+pub use view::*;
+
+pub(super) mod common;
