+++ web/app/package.json
@@ -7,17 +7,17 @@
   "scripts": {
     "build": "npm run build:css && npm run build:js",
     "build:css": "npx @tailwindcss/cli -i ./tailwind-custom.css -o ../public/css/tailwind.css --minify",
-    "build:js": "uglifyjs --compress --mangle -o ../public/js/core.min.js -- core.js && uglifyjs --compress --mangle -o ../public/js/toast.min.js -- toast.js && uglifyjs --compress --mangle -o ../public/js/wakelock.min.js -- wakelock.js && uglifyjs --compress --mangle -o ../public/js/json-highlighter.min.js -- json-highlighter.js"
+    "build:js": "uglifyjs --compress --mangle -o ../public/js/core.min.js -- core.js && uglifyjs --compress --mangle -o ../public/js/toast.min.js -- toast.js && uglifyjs --compress --mangle -o ../public/js/wakelock.min.js -- wakelock.js && uglifyjs --compress --mangle -o ../public/js/json-highlighter.min.js -- json-highlighter.js && uglifyjs --compress --mangle -o ../public/js/media.min.js -- media.js"
   },
   "license": "AGPL-3.0-or-later",
   "devDependencies": {
-    "@tailwindcss/cli": "^4.1.11",
+    "@tailwindcss/cli": "^4.1.13",
     "@tailwindcss/typography": "^0.5.16",
     "copyfiles": "^2.4.1",
-    "daisyui": "^5.0.54",
-    "esbuild": "^0.25.8",
+    "daisyui": "^5.1.7",
+    "esbuild": "^0.25.9",
     "mjml": "^4.15.3",
-    "tailwindcss": "^4.1.11",
+    "tailwindcss": "^4.1.13",
     "uglify-js": "^3.19.3"
   }
 }
