+++ web/public/css/app.css
@@ -118,3 +118,8 @@ dialog::backdrop,
     background-color: rgba(255, 0, 0, 0.2);
     color: #ff6666;
 }
+
+cropper-canvas {
+    width: 100%;
+    height: 100%;
+}
